3.0.0

+++ samples/server/petstore/jaxrs-cxf-cdi/src/main/java/io/swagger/api/impl/PetApiServiceImpl.java
@@ -25,7 +25,7 @@ public class PetApiServiceImpl implements PetApiService {
       return Response.ok().entity("magic!").build();
   }
       @Override
-      public Response deletePet(Integer petId, String apiKey, SecurityContext securityContext) {
+      public Response deletePet(Long petId, String apiKey, SecurityContext securityContext) {
       // do some magic!
       return Response.ok().entity("magic!").build();
   }
@@ -40,7 +40,7 @@ public class PetApiServiceImpl implements PetApiService {
       return Response.ok().entity("magic!").build();
   }
       @Override
-      public Response getPetById(Integer petId, SecurityContext securityContext) {
+      public Response getPetById(Long petId, SecurityContext securityContext) {
       // do some magic!
       return Response.ok().entity("magic!").build();
   }
@@ -50,12 +50,12 @@ public class PetApiServiceImpl implements PetApiService {
       return Response.ok().entity("magic!").build();
   }
       @Override
-      public Response updatePetWithForm(Integer petId, String name, String status, SecurityContext securityContext) {
+      public Response updatePetWithForm(Long petId, String name, String status, SecurityContext securityContext) {
       // do some magic!
       return Response.ok().entity("magic!").build();
   }
       @Override
-      public Response uploadFile(Integer petId, String additionalMetadata, InputStream fileInputStream, Attachment fileDetail, SecurityContext securityContext) {
+      public Response uploadFile(Long petId, String additionalMetadata, InputStream fileInputStream, Attachment fileDetail, SecurityContext securityContext) {
       // do some magic!
       return Response.ok().entity("magic!").build();
   }
